[Validator] Add a base AbstractLoader
